renamed:    change_dns.py -> DNS/change_dns.py 	new file:   DNS/show_dns.py 	modified:   install.sh 	new file:   menu.sh
